test(support): Extract query collecting into a helper module
