+++ test/support/query_collector.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module QueryCollector
+  attr_reader :queries
+
+  def setup
+    @queries = []
+    QueryNotifier.subscriber = ->(query) { @queries << query }
+  end
+
+  def teardown
+    QueryNotifier.subscriber = nil
+  end
+end
